+++ ios/lingdang/data/BXFoodProvider.m
@@ -13,7 +13,7 @@
 BCSINGLETON_IN_M(BXFoodProvider)
 
 - (void)addFoodWithName:(NSString*)name
-                  price:(int)price
+                  price:(float)price
                    shop:(BXShop*)shop
                 success:(void(^)(BXFood* food))sucBlock
                    fail:(void(^)(NSError* err))failBlock;
